Split MSP into protocol and serial sections.
