+++ src/main/io/serial_msp.h
@@ -57,32 +57,11 @@
  */
 
 // Each MSP port requires state and a receive buffer, revisit this default if someone needs more than 2 MSP ports.
+#include "io/msp_impl.h"
+
 #define MAX_MSP_PORT_COUNT 2
 
-typedef enum {
-    IDLE,
-    HEADER_START,
-    HEADER_M,
-    HEADER_ARROW,
-    HEADER_SIZE,
-    HEADER_CMD,
-    COMMAND_RECEIVED
-} mspState_e;
-
-#define MSP_PORT_INBUF_SIZE 64
-
-typedef struct mspPort_s {
-    serialPort_t *port; // null when port unused.
-    uint8_t offset;
-    uint8_t dataSize;
-    uint8_t checksum;
-    uint8_t indRX;
-    uint8_t inBuf[MSP_PORT_INBUF_SIZE];
-    mspState_e c_state;
-    uint8_t cmdMSP;
-} mspPort_t;
-
-void mspInit(void);
-void mspProcess(void);
+void mspSerialInit(void);
+void mspSerialProcess(void);
 void mspAllocateSerialPorts(void);
-void mspReleasePortIfAllocated(serialPort_t *serialPort);
+void mspReleaseSerialPortIfAllocated(serialPort_t *serialPort);
